on profile edit

+++ frontend/src/components/GroupChat/GroupChatIndex/Chat/GroupChat.jsx
@@ -1,11 +1,17 @@
 import "./GroupChat.css";
 import "./GroupChatBadge.css";
 import "../../../Profile/Profile.css";
-import { Link } from "react-router-dom";
+
+import { Link, useHistory } from "react-router-dom";
 import { jwtFetch } from "../../../../store/jwt";
-import { useEffect } from "react";
+import { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { leaveChat } from "../../../../store/chats";
+import {
+  addError,
+  addPassChat,
+  clearChatErrors,
+} from "../../../../store/chats";
+import { ReactComponent as Lock } from "./SVG/icons8-lock.svg";
 
 function GroupChat({
   chatData,
@@ -16,12 +22,20 @@ function GroupChat({
   setConfirmDelete,
   socket,
 }) {
+  const chatError = useSelector((state) => state.errors.chat);
+
+  const [openChatPassword, setOpenChatPassword] = useState(false);
+  const [password, setPassword] = useState("");
+  const history = useHistory();
+  const dispatch = useDispatch();
+
   const handleOnClickDelete = () => {
     setConfirmDelete(false);
-
-    jwtFetch(`/api/groupchats/image/delete/${chatData.logo}`, {
-      method: "DELETE",
-    }).then((res) => console.log(res));
+    if (chatData.logo !== "c-logo.png") {
+      jwtFetch(`/api/groupchats/image/delete/${chatData.logo}`, {
+        method: "DELETE",
+      }).then((res) => console.log(res));
+    }
 
     jwtFetch(`/api/groupchats/delete/${chatData._id}`, {
       method: "DELETE",
@@ -35,10 +49,32 @@ function GroupChat({
     setAllChats([...tempChats]);
   };
 
+  const handlePasswordCheck = (e) => {
+    if (chatData.password) {
+      e.preventDefault();
+      setOpenChatPassword(chatData._id);
+    }
+  };
+
+  const handleEnterChatRoomWithPassword = (e) => {
+    if (password === chatData.password) {
+      dispatch(clearChatErrors());
+      dispatch(addPassChat(chatData._id));
+      history.push(`/groupchats/${chatData._id}`);
+    } else {
+      dispatch(clearChatErrors());
+      dispatch(addError({ invalid_password: "Incorrect Password" }));
+    }
+  };
+
   return (
     <>
       <div className="group-chat-container">
-        <Link className="group-chat" to={`/groupchats/${chatData._id}`}>
+        <Link
+          className="group-chat"
+          to={`/groupchats/${chatData._id}`}
+          onClick={handlePasswordCheck}
+        >
           <div className="top-groupchat">
             <img
               src={`http://localhost:5000/api/groupchats/image/${chatData.logo}`}
@@ -52,8 +88,16 @@ function GroupChat({
             <p className="description-body">{chatData.description}</p>
           </div>
           <div className="bottom-groupchat">
-            <div className="badge mini">Animals</div>
-            <div className="badge mini">Pictures</div>
+            {chatData.password && (
+              <p id="password-protected">
+                <Lock className="password-lock-svg" />
+              </p>
+            )}
+            <div className="badges">
+              {chatData?.tags.map((tag) => {
+                return <div className="badge mini">{tag.text}</div>;
+              })}
+            </div>
           </div>
         </Link>
         {chatData.owner === user._id && (
@@ -75,6 +119,36 @@ function GroupChat({
                 Confirm
               </span>
               <span id="no" onClick={(e) => setConfirmDelete(false)}>
+                Cancel
+              </span>
+            </div>
+          </div>
+        </>
+      )}
+      {openChatPassword === chatData._id && (
+        <>
+          <div className="blur-background"></div>
+          <div className="confirm-delete-modal">
+            <p id="warning">Enter Password for "{chatData.title}".</p>
+            {chatError?.invalid_password}
+            <input
+              className="password-chat-input"
+              type="password"
+              value={password}
+              onChange={(e) => setPassword(e.target.value)}
+            />
+            <div className="options-confirm">
+              <span id="yes" onClick={handleEnterChatRoomWithPassword}>
+                Confirm
+              </span>
+              <span
+                id="no"
+                onClick={(e) => {
+                  dispatch(clearChatErrors());
+                  setOpenChatPassword(false);
+                  setPassword("");
+                }}
+              >
                 Cancel
               </span>
             </div>
